model-service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
onnxruntime==1.15.1
Pillow==9.5.0
python-multipart==0.0.6
